router also moved back to class
