+++ linuxConsole.vcxproj
@@ -88,6 +88,7 @@
     <ClInclude Include="lght.h" />
     <ClInclude Include="lightDriver.h" />
     <ClInclude Include="router.h" />
+    <ClInclude Include="safeItr.h" />
     <ClInclude Include="safePtr.h" />
   </ItemGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
